% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chord_segment.R
\name{chord_segment}
\alias{chord_segment}
\title{chord_segment}
\usage{
chord_segment(out_start, out_end, in_start, in_end)
}
\arguments{
\item{out_start}{point on circle (value in [0, 1])}

\item{out_end}{point on circle (value in [0, 1])}

\item{in_start}{point on circle (value in [0, 1])}

\item{in_end}{point on circle (value in [0, 1])}
}
\description{
Creates x,y boundaries for chord segments of the chord diagram
  circle. \cr start and end arguments are in the interval [0, 1] and are
  mapped to [0, 2 * pi]
}

