% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/circly.R
\docType{package}
\name{circly-package}
\alias{circly-package}
\title{Package for chord diagrams}
\description{
Package for chord diagrams
}
\examples{
circly(matrix(1:9, ncol = 3), matrix(9:1, ncol = 3))
}
\references{
Useful Stack Overflow comment from Vincent Zoonekynd: \cr
  http://stackoverflow.com/questions/14599150/chord-diagram-in-r
}

