% Generated by roxygen2 (4.0.2): do not edit by hand
\name{circly}
\alias{circly}
\title{circly}
\usage{
circly(M1, M2 = M1, color_column = rep(rgb(0, 0, 0), ncol(M1)),
  color_row = rep(rgb(0, 0, 0), nrow(M1)), gap_width = 0.005, cex = 1,
  label_orientation = "radial", ROI = 1)
}
\arguments{
\item{M1}{n x m matrix. Columns specify output segments and are placed at the
bottom of the circle. Rows specify input segments and are placed at the
top.}

\item{M2}{n x m matrix. If not specified, M1 is used.}

\item{color_column}{Colours for lower segments and their chords.}

\item{color_row}{Colours for top segments.}

\item{gap_width}{Proportion of circle used for gaps.}

\item{ROI}{Value indicating the proportion that each hemisphere should cover.
e.g. value of 2 makes top hemisphere cover 2/3 of the circle}
}
\description{
circly
}

