% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outer_segment}
\alias{outer_segment}
\title{outer_segment}
\usage{
outer_segment(start, end, inner_r, outer_r)
}
\arguments{
\item{start}{point on circle (value in [0, 1])}

\item{end}{point on circle (value in [0, 1])}

\item{inner_r}{inner radius of segment}

\item{outer_r}{outer radius of segment}
}
\description{
Creates x,y boundaries for outer segments of the chord diagram
  circle. \cr start and end arguments are in the interval [0, 1] and are
  mapped to [0, 2 * pi]
}

