% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outer_segment.R
\name{outer_segments}
\alias{outer_segments}
\title{outer_segments
calls outer_segment for a data.frame that lists numerous outer segments}
\usage{
outer_segments(segment)
}
\arguments{
\item{segment}{A data.frame with Start, End, Label, and Direction columns}
}
\description{
outer_segments
calls outer_segment for a data.frame that lists numerous outer segments
}

