% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chord_spacing.R
\name{respace}
\alias{respace}
\title{Function for spacing chords along the arc of a circle.}
\usage{
respace(x, id, direction, gap_width = 0.005, ROI = 1)
}
\arguments{
\item{x}{Numeric vector. Size of element in x corresponds to the width of the
corresponding chord.}

\item{id}{Character or factor vector indicating how values of x are grouped.
Most notably used to place gaps. Should be arranged so that groups appear
in sequence.}

\item{direction}{Either 'out' or 'in'.}

\item{gap_width}{Space between outer segments of chord diagram default is
0.005}

\item{ROI}{Rescaling of input or output hemispheres. Default is 1 (both input
and output hemispheres)}
}
\description{
Function for spacing chords along the arc of a circle.
}

