\name{circly-package}
\alias{circly-package}
\alias{circly}
\docType{package}
\title{
Creates chord diagrams
}
\description{
Creates chord diagrams
}
\details{
\tabular{ll}{
Package: \tab circly\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-07\cr
License: \tab \cr
}

}
\author{
Wojciech (Wojtek) Kostelecki

Maintainer: <wkostelecki@gmail.com>
}
\references{
Useful Stack Overflow comment from Vincent Zoonekynd:
http://stackoverflow.com/questions/14599150/chord-diagram-in-r
}

\keyword{ package }
\seealso{

}
\examples{
circly(matrix(1:9, ncol = 3), matrix(9:1, ncol = 3))
}
